+++ src/main/java/io/github/dataforseo/client/model/RankedSerpElement.java
@@ -138,9 +138,9 @@ public class RankedSerpElement  {
 
   public static final String SERIALIZED_NAME_SE_RESULTS_COUNT = "se_results_count";
   @SerializedName(SERIALIZED_NAME_SE_RESULTS_COUNT)
-  private Integer seResultsCount;
+  private Long seResultsCount;
 
-  public RankedSerpElement seResultsCount(Integer seResultsCount) {
+  public RankedSerpElement seResultsCount(Long seResultsCount) {
     this.seResultsCount = seResultsCount;
     return this;
   }
@@ -150,11 +150,11 @@ public class RankedSerpElement  {
    * @return seResultsCount
    */
   @javax.annotation.Nullable
-  public Integer getSeResultsCount() {
+  public Long getSeResultsCount() {
     return seResultsCount;
   }
 
-  public void setSeResultsCount(Integer seResultsCount) {
+  public void setSeResultsCount(Long seResultsCount) {
     this.seResultsCount = seResultsCount;
   }
